feat(ui): created a starter page
